requirements.txt
opentelemetry-api==1.11.0
opentelemetry-exporter-otlp-proto-http==1.11.0
opentelemetry-exporter-datadog
opentelemetry-instrumentation